Change dependency configuration and Dockerfile.
The dependecy of "db" is now pulled from ../db, when working locally.
However, when this crate is installed from crates.io the verison 0.1.0
will be used.
The Dockerfile will now only build releases and the Dockerfile.dev will
be used to build development container, but currently this won't be
possible as the container is not configured to use the local db which is
no longer a submodule.